pyproject.toml
[project]
name = "colibrilocal"
version = "0.1.0"
description = "Local PySpark + Delta Lake medallion pipeline for wind turbine power data."
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "delta-spark==3.2.0",
    "pyspark==3.5.3",
]

[project.optional-dependencies]
dev = [
    "pytest==9.0.2",
]

[project.scripts]
colibri-wt-power = "src.pipelines.pl_wt_power:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

